requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
python-multipart
reportlab
apscheduler
langchain-google-genai
langchain
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
